delete new plugin
